kernelci.config: rename .lab as .runtime
Rename the kernelci.config.lab module as kernelci.config.runtime.
Update all other parts of the code using it accordingly.
